refactor: remove outdated integration tests and update test structure
