--- vscode-extension/test/integration/extension.test.ts
@@ -1,46 +1,0 @@
-import * as assert from 'assert';
-
-// You can import and use all API from the 'vscode' module
-// as well as import your extension to test it
-import * as vscode from 'vscode';
-// import * as myExtension from '../../extension';
-
-suite('Extension Test Suite', () => {
-	vscode.window.showInformationMessage('Start all tests.');
-
-	test('Sample test', () => {
-		assert.strictEqual(-1, [1, 2, 3].indexOf(5));
-		assert.strictEqual(-1, [1, 2, 3].indexOf(0));
-	});
-
-	test('Extension should be present', () => {
-		const extension = vscode.extensions.getExtension('RobBos.ai-engineering-fluency');
-		assert.ok(extension, 'Extension should be installed');
-	});
-
-	test('Commands should be registered', async () => {
-		const extension = vscode.extensions.getExtension('RobBos.ai-engineering-fluency');
-		if (extension && !extension.isActive) {
-			await extension.activate();
-		}
-
-		const commands = await vscode.commands.getCommands(true);
-		
-		const expectedCommands = [
-			'aiEngineeringFluency.refresh',
-			'aiEngineeringFluency.showDetails',
-			'aiEngineeringFluency.showChart',
-			'aiEngineeringFluency.showMaturity',
-			'aiEngineeringFluency.showFluencyLevelViewer',
-			'aiEngineeringFluency.runLocalViewRegression',
-			'aiEngineeringFluency.generateDiagnosticReport'
-		];
-
-		for (const expectedCommand of expectedCommands) {
-			assert.ok(
-				commands.includes(expectedCommand),
-				`Command ${expectedCommand} should be registered`
-			);
-		}
-	});
-});
