+++ vscode-extension/test/unit/vscode-shim-register.ts
@@ -271,10 +271,21 @@ function attachMock(target: any): void {
 	target.extensions = target.extensions ?? {};
 	target.extensions.getExtension = target.extensions.getExtension ?? ((id: string) => {
 		if (id === 'RobBos.ai-engineering-fluency' || id === 'RobBos.copilot-token-tracker') {
-			return { packageJSON: { version: '0.0.0-test' } };
+			return { isActive: true, packageJSON: { version: '0.0.0-test' }, activate: async () => {} };
 		}
 		return state.extensions[id] as any;
 	});
+
+	target.commands = target.commands ?? {};
+	target.commands.getCommands = target.commands.getCommands ?? (async (_filterInternal?: boolean) => [
+		'aiEngineeringFluency.refresh',
+		'aiEngineeringFluency.showDetails',
+		'aiEngineeringFluency.showChart',
+		'aiEngineeringFluency.showMaturity',
+		'aiEngineeringFluency.showFluencyLevelViewer',
+		'aiEngineeringFluency.runLocalViewRegression',
+		'aiEngineeringFluency.generateDiagnosticReport'
+	]);
 }
 
 const vscodeStub: any = {};
